+++ selector/parser/parse_group_by.go
@@ -1,0 +1,35 @@
+package parser
+
+import (
+	"github.com/tomwright/dasel/v3/selector/ast"
+	"github.com/tomwright/dasel/v3/selector/lexer"
+)
+
+func parseGroupBy(p *Parser) (ast.Expr, error) {
+	if err := p.expect(lexer.GroupBy); err != nil {
+		return nil, err
+	}
+	p.advance()
+
+	expr, err := p.parseExpressionsFromTo(
+		lexer.OpenParen,
+		lexer.CloseParen,
+		[]lexer.TokenKind{},
+		true,
+		bpDefault,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	groupByExpr := ast.GroupByExpr{
+		Expr: expr,
+	}
+
+	res, err := parseFollowingSymbol(p, groupByExpr)
+	if err != nil {
+		return nil, err
+	}
+
+	return res, nil
+}